requirements.txt
Flask
pandas
google-generativeai
sqlalchemy
rapidfuzz
scikit-learn
numpy
pymysql
